modified:   CMakeLists.txt 	modified:   src/chatbot.cpp 	modified:   src/chatbot.h 	modified:   src/chatgui.cpp 	modified:   src/chatgui.h 	modified:   src/chatlogic.cpp 	modified:   src/graphnode.h

+++ src/chatlogic.cpp
@@ -98,7 +98,7 @@ void ChatLogic::LoadAnswerGraphFromFile(std::string filename)
                 std::string tokenStr = lineStr.substr(posTokenFront + 1, posTokenBack - 1);
 
                 // extract token type and info
-                int posTokenInfo = tokenStr.find(":");
+                size_t posTokenInfo = tokenStr.find(":");
                 if (posTokenInfo != std::string::npos)
                 {
                     std::string tokenType = tokenStr.substr(0, posTokenInfo);
